Add files via upload
This is a knit Word document I produced using my R script. It features the visuals I produced based on my analysis of my peace agreement data and an initial coding of full civil war peace agreements. Unfortunately, I was not able to get the document to knit without printing out the code in the Word document, and so I deleted that manually.