Add files via upload
This is the full Uppsala Conflict Data Program (UCDP) dataframe. As demonstrated in my R script, I initially tried to work with this data to get to the civil wars I wanted to analyze, but found a more direct route by downloading the UCDP dataframe that is also uploaded to this repo.