Add files via upload
This is a knit Word document that features the output of my R script, as well as a commentary at what I attempted to do, and what I found, at each stage of the project. I struggled to get the knitted document to not include my code, so I manually deleted that in the attached.