Add files via upload
This is a knit PDF that features the output of my R script, as well as a commentary at what I attempted to do, and what I found, at each stage of the project. I struggled to get this to knit as a PDF, so I manually converted the knitted Word document into the attached.